refactor: extract OverlayClassListProxy to shared package
