+++ vaadin-dialog-flow-parent/vaadin-dialog-flow/src/main/java/com/vaadin/flow/component/dialog/Dialog.java
@@ -16,10 +16,8 @@
 package com.vaadin.flow.component.dialog;
 
 import java.io.Serializable;
-import java.util.AbstractSet;
 import java.util.Collection;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
@@ -41,6 +39,7 @@ import com.vaadin.flow.component.UI;
 import com.vaadin.flow.component.dependency.JsModule;
 import com.vaadin.flow.component.dependency.NpmPackage;
 import com.vaadin.flow.component.shared.HasThemeVariant;
+import com.vaadin.flow.component.shared.InternalOverlayClassListProxy;
 import com.vaadin.flow.dom.ClassList;
 import com.vaadin.flow.dom.Element;
 import com.vaadin.flow.dom.ElementConstants;
@@ -993,7 +992,7 @@ public class Dialog extends Component implements HasComponents, HasSize,
 
     @Override
     public ClassList getClassNames() {
-        return new OverlayClassListProxy(this);
+        return new InternalOverlayClassListProxy(this);
     }
 
     /**
@@ -1004,62 +1003,5 @@ public class Dialog extends Component implements HasComponents, HasSize,
     public Style getStyle() {
         throw new UnsupportedOperationException(
                 "Dialog does not support adding styles to overlay");
-    }
-
-    static class OverlayClassListProxy extends AbstractSet<String>
-            implements ClassList {
-        private final HasStyle hasStyle;
-        private final ClassList classList;
-
-        public OverlayClassListProxy(HasStyle hasStyle) {
-            this.hasStyle = hasStyle;
-            this.classList = hasStyle.getElement().getClassList();
-        }
-
-        private void updateOverlayClass() {
-            hasStyle.getElement().setProperty("overlayClass",
-                    hasStyle.getClassName());
-        }
-
-        @Override
-        public Iterator<String> iterator() {
-            return new IteratorProxy(classList.iterator());
-        }
-
-        @Override
-        public int size() {
-            return classList.size();
-        }
-
-        @Override
-        public boolean add(String s) {
-            boolean result = classList.add(s);
-            updateOverlayClass();
-            return result;
-        }
-
-        private class IteratorProxy implements Iterator<String> {
-            private final Iterator<String> iterator;
-
-            public IteratorProxy(Iterator<String> iterator) {
-                this.iterator = iterator;
-            }
-
-            @Override
-            public boolean hasNext() {
-                return iterator.hasNext();
-            }
-
-            @Override
-            public String next() {
-                return iterator.next();
-            }
-
-            @Override
-            public void remove() {
-                iterator.remove();
-                updateOverlayClass();
-            }
-        }
     }
 }
